# In-place call dispatch and reduced memory allocations