--- a/interpreter/interpreter.go
+++ b/interpreter/interpreter.go
@@ -28,7 +28,7 @@ import (
 // Interpreter generates a new Interpretable from a Program.
 type Interpreter interface {
 	// NewInterpretable returns an Interpretable from a Program.
-	NewInterpretable(program Program) Interpretable
+	NewInterpretable(program *Program) Interpretable
 }
 
 // Interpretable can accept a given Activation and produce a value along with
@@ -66,28 +66,36 @@ func NewStandardInterpreter(packager packages.Packager,
 	return NewInterpreter(dispatcher, packager, typeProvider)
 }
 
-func (i *exprInterpreter) NewInterpretable(program Program) Interpretable {
+func (i *exprInterpreter) NewInterpretable(program *Program) Interpretable {
 	// program needs to be pruned with the TypeProvider
 	evalState := NewEvalState(program.MaxInstructionID() + 1)
-	program.Init(i.dispatcher, evalState)
+	intr := &exprInterpreter{
+		dispatcher:   i.dispatcher.Init(evalState),
+		packager:     i.packager,
+		typeProvider: i.typeProvider}
+	program.Init(evalState)
 	return &exprInterpretable{
-		interpreter: i,
+		interpreter: intr,
 		program:     program,
 		state:       evalState}
 }
 
 type exprInterpretable struct {
 	interpreter *exprInterpreter
-	program     Program
+	program     *Program
 	state       MutableEvalState
 }
 
 func (i *exprInterpretable) Eval(activation Activation) (ref.Value, EvalState) {
 	// register machine-like evaluation of the program with the given activation.
 	currActivation := activation
-	stepper := i.program.Begin()
+	// program counter
+	pc := 0
+	end := len(i.program.Instructions)
+	steps := i.program.Instructions
 	var resultID int64
-	for step, hasNext := stepper.Next(); hasNext; step, hasNext = stepper.Next() {
+	for pc < end {
+		step := steps[pc]
 		resultID = step.GetID()
 		switch step.(type) {
 		case *IdentExpr:
@@ -108,11 +116,14 @@ func (i *exprInterpretable) Eval(activation Activation) (ref.Value, EvalState) {
 		case *JumpInst:
 			jmpExpr := step.(*JumpInst)
 			if jmpExpr.OnCondition(i.state) {
-				if !stepper.JumpCount(jmpExpr.Count) {
+				jmpPc := pc + jmpExpr.Count
+				if jmpPc < 0 || jmpPc >= end {
 					// TODO: Error, the jump count should never exceed the
 					// program length.
 					panic("jumped too far")
 				}
+				pc = jmpPc
+				continue
 			}
 			// Special instructions for modifying the activation stack
 		case *PushScopeInst:
@@ -129,6 +140,7 @@ func (i *exprInterpretable) Eval(activation Activation) (ref.Value, EvalState) {
 		case *PopScopeInst:
 			currActivation = currActivation.Parent()
 		}
+		pc++
 	}
 	result := i.value(resultID)
 	if result == nil {
@@ -232,21 +244,17 @@ func (i *exprInterpretable) resolveUnknown(unknown types.Unknown,
 }
 
 func (i *exprInterpretable) evalCall(callExpr *CallExpr, currActivation Activation) {
-	argVals := make([]ref.Value, len(callExpr.Args), len(callExpr.Args))
-	for idx, argID := range callExpr.Args {
-		argVals[idx] = i.value(argID)
-		if callExpr.Strict && types.IsUnknownOrError(argVals[idx]) {
-			i.setValue(callExpr.GetID(), argVals[idx])
-			return
+	if callExpr.Strict {
+		for _, argID := range callExpr.Args {
+			argVal := i.value(argID)
+			argType := argVal.Type()
+			if types.IsUnknownOrError(argType) {
+				i.setValue(callExpr.GetID(), argVal)
+				return
+			}
 		}
 	}
-	ctx := &CallContext{
-		call:       callExpr,
-		activation: currActivation,
-		args:       argVals,
-		metadata:   i.program.Metadata()}
-	result := i.interpreter.dispatcher.Dispatch(ctx)
-	i.setValue(callExpr.GetID(), result)
+	i.interpreter.dispatcher.Dispatch(callExpr)
 }
 
 func (i *exprInterpretable) evalCreateList(listExpr *CreateListExpr) {