… interpreter/dispatcher_test.go
… interpreter/interpreter_test.go
… interpreter/program_test.go