The call dispatch has been made in-place meaning that the `Dispatcher` now modifies
loads / stores state directly for function calls rather than create a new `CallContext` and
arg list for each invocation. This reduces allocations and improves performance significantly.

Trivial allocations related to tracking the `Program` instruction counter and checking type
names have been scrubbed.

Below is a table of the before and after benchmarks and deltas:
```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkInterpreter_ConditionalExpr-12                              2958          1999          -32.42%
BenchmarkInterpreter_EqualsCall-12                                   163           102           -37.42%
BenchmarkInterpreter_EqualsDispatch-12                               743           186           -74.97%
BenchmarkInterpreter_EqualInstructions-12                            1194          486           -59.30%
BenchmarkInterpreter_ComprehensionExpr-12                            5422          3898          -28.11%
BenchmarkInterpreter_ComprehensionExprWithInput-12                   3467          2375          -31.50%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/ok_1st-12        1383          597           -56.83%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/ok_2nd-12        3693          1742          -52.83%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/not_found-12     3497          1728          -50.59%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/false_2nd-12     724           307           -57.60%

benchmark                                                            old allocs     new allocs     delta
BenchmarkInterpreter_ConditionalExpr-12                              17             12             -29.41%
BenchmarkInterpreter_EqualsCall-12                                   1              0              -100.00%
BenchmarkInterpreter_EqualsDispatch-12                               5              0              -100.00%
BenchmarkInterpreter_EqualInstructions-12                            8              0              -100.00%
BenchmarkInterpreter_ComprehensionExpr-12                            22             13             -40.91%
BenchmarkInterpreter_ComprehensionExprWithInput-12                   17             10             -41.18%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/ok_1st-12        7              1              -85.71%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/ok_2nd-12        16             5              -68.75%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/not_found-12     16             5              -68.75%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/false_2nd-12     3              0              -100.00%

benchmark                                                            old bytes     new bytes     delta
BenchmarkInterpreter_ConditionalExpr-12                              416           224           -46.15%
BenchmarkInterpreter_EqualsCall-12                                   16            0             -100.00%
BenchmarkInterpreter_EqualsDispatch-12                               192           0             -100.00%
BenchmarkInterpreter_EqualInstructions-12                            168           0             -100.00%
BenchmarkInterpreter_ComprehensionExpr-12                            1056          720           -31.82%
BenchmarkInterpreter_ComprehensionExprWithInput-12                   800           544           -32.00%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/ok_1st-12        224           8             -96.43%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/ok_2nd-12        528           104           -80.30%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/not_found-12     528           104           -80.30%
BenchmarkInterpreter_CanonicalExpressions/ExprBench/false_2nd-12     112           0             -100.00%
```

Further improvements can be made to reduce the overhead within the `Dispatcher` as well as the
overhead between the `Interpreter` and `Dispatcher` as seen in the `EqualsCall`, `EqualsDispatch`,
and `EqualsInstructions` `ns/op` times.