@TristonianJones 2019-01-09
@kyessenov the `CanonicalExpressions` line up with some of the Istio CEXL tests that you put together if you are interested.